Implement String in pure Axion, add HashSet[T], bool as i1
- Move String from ~20 codegen intrinsics to pure Axion implementation
  using FFI (malloc/memcpy) + Ptr[u8] operations
- Add str.as_ptr() and str_from_raw() bridge intrinsics for str<->Ptr[u8]
- Add Constructor codegen support (declare + compile static methods)
- Change bool LLVM type from i8 to i1, remove all i1->i8 coercions
- Implement HashSet[T] with open-addressing hash table (djb2, linear probing)

+++ packages/axion-codegen/src/tests.rs
@@ -2534,7 +2534,7 @@ fn main() -> i64
 #[test]
 fn compile_hashset_new() {
     let src = "\
-fn main() -> i64
+fn main() -> usize
     let s = HashSet[i64].new()
     s.len()
 ";
